+++ src/main/java/com/xqfitness/readservice/repository/MuscleGroupRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MuscleGroupRepository extends JpaRepository<MuscleGroup, Long> {
+public interface MuscleGroupRepository extends JpaRepository<MuscleGroup, Integer> {
 }
